raw/requirements.txt
pandas==1.3.3
kafka-python==2.0.2
watchdog==2.1.6
